requirements.txt
django==2.0
django-extensions
django-cors-headers
djangorestframework
djangorestframework-jwt
psycopg2
django-polymorphic
django-datatable-view